Minor typo in Hero class
compile and copy paste reunion-core.jar

+++ L2J_ReunionProject_Core/java/l2r/gameserver/model/entity/Hero.java
@@ -326,7 +326,7 @@ public class Hero
 			int charTwoClass;
 			int winner;
 			long start;
-			int time;
+			long time;
 			int classed;
 			while (rset.next())
 			{
@@ -336,7 +336,7 @@ public class Hero
 				charTwoClass = rset.getInt("charTwoClass");
 				winner = rset.getInt("winner");
 				start = rset.getLong("start");
-				time = rset.getInt("time");
+				time = rset.getLong("time");
 				classed = rset.getInt("classed");
 				
 				if (charId == charOneId)
